Updated to latest version of proposal
- Refactors MessageInterface to only provide getters.
- MessageInterface now defines getBody() to require that it return a
  StreamableInterface instance.
- Removes Request and Response interfaces
- Provides server-side interfaces:
  - IncomingRequestInterface, which provides accessors for HTTP
    properties and environment-specific items ($_SERVER, $_GET, $_POST,
    $_FILES, $_COOKIE, etc), and support for mutable "attributes".
  - OutgoingResponseInterface, which provides both accessors and
    mutators for all HTTP properties.
- Provides client-side interfaces:
  - OutgoingRequestInterface, which provides accessors and mutators for
    all HTTP properties.
  - IncomingResponseInterface, which provides accessors for HTTP
    properties.
- StreamableInterface removes attach().
