+++ src/MessageInterface.php
@@ -22,34 +22,11 @@ interface MessageInterface
     public function getProtocolVersion();
 
     /**
-     * Set the HTTP protocol version.
-     *
-     * The version string MUST contain only the HTTP version number (e.g.,
-     * "1.1", "1.0").
-     *
-     * @param string $version HTTP protocol version
-     * @return void
-     */
-    public function setProtocolVersion($version);
-
-    /**
      * Gets the body of the message.
      *
      * @return StreamableInterface|null Returns the body, or null if not set.
      */
     public function getBody();
-
-    /**
-     * Sets the body of the message.
-     *
-     * The body MUST be a StreamableInterface object. Setting the body to null MUST
-     * remove the existing body.
-     *
-     * @param StreamableInterface|null $body Body.
-     * @return void
-     * @throws \InvalidArgumentException When the body is not valid.
-     */
-    public function setBody(StreamableInterface $body = null);
 
     /**
      * Gets all message headers.
@@ -85,11 +62,14 @@ interface MessageInterface
     public function hasHeader($header);
 
     /**
-     * Retrieve a header by the given case-insensitive name as a string.
+     * Retrieve a header by the given case-insensitive name, as a string.
      *
      * This method returns all of the header values of the given
      * case-insensitive header name as a string concatenated together using
      * a comma.
+     *
+     * NOTE: Not all header values may be appropriately represented using
+     * comma concatenation.
      *
      * @param string $header Case-insensitive header name.
      * @return string
@@ -103,39 +83,4 @@ interface MessageInterface
      * @return string[]
      */
     public function getHeaderAsArray($header);
-
-    /**
-     * Sets a header, replacing any existing values of any headers with the
-     * same case-insensitive name.
-     *
-     * The header name is case-insensitive. The header values MUST be a string
-     * or an array of strings.
-     *
-     * @param string $header Header name
-     * @param string|string[] $value Header value(s).
-     * @return void
-     * @throws \InvalidArgumentException for invalid header names or values.
-     */
-    public function setHeader($header, $value);
-
-    /**
-     * Appends a header value for the specified header.
-     *
-     * Existing values for the specified header will be maintained. The new
-     * value(s) will be appended to the existing list.
-     *
-     * @param string $header Header name to add
-     * @param string|string[] $value Header value(s).
-     * @return void
-     * @throws \InvalidArgumentException for invalid header names or values.
-     */
-    public function addHeader($header, $value);
-
-    /**
-     * Remove a specific header by case-insensitive name.
-     *
-     * @param string $header HTTP header to remove
-     * @return void
-     */
-    public function removeHeader($header);
 }
