+++ src/StreamableInterface.php
@@ -40,25 +40,6 @@ interface StreamableInterface
     public function detach();
 
     /**
-     * Replaces the underlying stream resource with the provided stream.
-     *
-     * Use this method to replace the underlying stream with another; as an
-     * example, in server-side code, if you decide to return a file, you
-     * would replace the original content-oriented stream with the file
-     * stream.
-     *
-     * Any internal state such as caching of cursor position should be reset
-     * when attach() is called, as the stream has changed.
-     *
-     * @param string|resource $stream The underlying stream. String values
-     *                                SHOULD be used to create a stream
-     *                                resource.
-     * @return void
-     * @throws \InvalidArgumentException For invalid $stream arguments.
-     */
-    public function attach($stream);
-
-    /**
      * Get the size of the stream if known
      *
      * @return int|null Returns the size in bytes if known, or null if unknown.
